bug:0000003 msg:Controllar les FichaUA que no siguin null quan feim la baixa d'una administració Remota
git-svn-id: https://svn.code.sf.net/p/rolsac/code/rolsac-core/trunk@107 58ef2255-6f3f-491f-b93b-e266552ec6fa

+++ moduls/persistence/src/org/ibit/rol/sac/persistence/ejb/FichaRemotaFacadeEJB.java
@@ -352,10 +352,12 @@ public abstract class FichaRemotaFacadeEJB extends HibernateEJB {
 			if (seccion != null && uaRemota != null && fichaRemota != null) {
 				final FichaUA fichaUA = RemotoUtils.recogerFichaUA(session,
 						uaRemota.getId(), fichaRemota.getId(), seccion.getId());
-				seccion.removeFichaUA(fichaUA);
-				uaRemota.removeFichaUA(fichaUA);
-				fichaRemota.removeFichaUA(fichaUA);
-				session.delete(fichaUA);
+                if(fichaUA!=null){
+				    seccion.removeFichaUA(fichaUA);
+				    uaRemota.removeFichaUA(fichaUA);
+				    fichaRemota.removeFichaUA(fichaUA);
+				    session.delete(fichaUA);
+                }
 			}
 			session.flush();
 		} catch (final HibernateException e) {
